requirements.txt
Flask
pillow
tensorflow
keras